services/agent-orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
redis==5.0.1
httpx==0.25.2
prometheus-client==0.19.0
temporalio==1.5.0
langgraph==0.0.21
langchain==0.1.0
langchain-core==0.1.0
nats-py==2.6.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-exporter-otlp==1.22.0
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
ruff==0.1.8
mypy==1.7.1
